Fix uploading GPX files to subdirectories.
SkyDrive REST API doesn't like filenames with a "/" in them. You need to
look up the containing folder ID and use that...

For deeply nested paths, or just a lot of files in subfolders, this is
going to incur a lot of HTTP GET requests. A global shared cache of
mappings from path to SkyDrive folder IDs would be nice to reduce this,
so we only have to discover paths once. Adding as a TODO for later.